move build result endpoint to public
