+++ src/test/java/de/tum/in/www1/artemis/programmingexercise/ProgrammingExerciseResultTestService.java
@@ -174,7 +174,7 @@ public class ProgrammingExerciseResultTestService {
 
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.add("Authorization", ARTEMIS_AUTHENTICATION_TOKEN_VALUE);
-        request.postWithoutLocation("/api/programming-exercises/new-result", alteredObj, HttpStatus.OK, httpHeaders);
+        request.postWithoutLocation("/api/public/programming-exercises/new-result", alteredObj, HttpStatus.OK, httpHeaders);
     }
 
     private ProgrammingExerciseTestCase createTest(String testName, long testId, ProgrammingExerciseTestCaseType testCaseType) {
