+++ src/test/java/de/tum/in/www1/artemis/programmingexercise/ProgrammingSubmissionAndResultGitlabJenkinsIntegrationTest.java
@@ -431,7 +431,7 @@ class ProgrammingSubmissionAndResultGitlabJenkinsIntegrationTest extends Abstrac
 
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.add("Authorization", ARTEMIS_AUTHENTICATION_TOKEN_VALUE);
-        request.postWithoutLocation("/api/programming-exercises/new-result", alteredObj, status, httpHeaders);
+        request.postWithoutLocation("/api/public/programming-exercises/new-result", alteredObj, status, httpHeaders);
     }
 
     private TestResultsDTO createJenkinsNewResultNotification(String projectKey, String loginName, ProgrammingLanguage programmingLanguage, List<String> successfulTests,
